avoid rounding errors.
It turns out that on some web browsers (my version of firefox for
example), rounding numbers doesn't work as predictably as we would
expct.

For example:

Math.round(99.995 * 100)/100 => 100 (as expected)

But...

Math.round(77.475 * 100)/100 => 77.47 (what?!?)

The number is properly rounded by PHP on the backend. But displayed to
the user incorrectly.

This causes a visual discrepancy for the user, but also problems with
Stripe since one amount is used for the payment intent and then, when
the payment intent is applied on the backend, it's different and you
get the error:

payment_intent_unexpected_state

This PaymentIntent's amount could not be updated because it has a status
of requires_capture. You may only update the amount of a PaymentIntent
with one of the following statuses: requires_payment_method,
requires_confirmation, requires_action.